+++ lib/shrine/plugins/activerecord.rb
@@ -131,7 +131,7 @@ class Shrine
 
         def activerecord_json_column?
           return false unless record.is_a?(ActiveRecord::Base)
-          return false unless column = record.class.columns_hash["#{name}_data"]
+          return false unless column = record.class.columns_hash[data_attribute.to_s]
 
           [:json, :jsonb].include?(column.type)
         end